Basic 64-bit support + Fix crash when saving save state

+++ emu-ex-plus-alpha/GBA.emu/src/vbam/common/Patch.cpp
@@ -91,7 +91,7 @@ static s64 readVarPtr(FILE *f)
 static uLong computePatchCRC(FILE *f, unsigned int size)
 {
   Bytef buf[4096];
-  long readed;
+  size_t readed;
 
   uLong crc = crc32(0L, Z_NULL, 0);
   do {
